requirements.txt
aiohttp
discord
discord.py
python-dotenv
pytz
